Split out option into multiple submodules
